feat: add useParent & useChildren
