+++ src/packages/__VUE/form/types.ts
@@ -1,5 +1,7 @@
 import { FormItemRule } from '../formitem/types';
 
+export const FORM_KEY = Symbol('nut-form');
+
 export type FormRules = {
   [key: string]: FormItemRule[];
 };
