+++ src/packages/utils/useRelation/useParent.ts
@@ -1,0 +1,14 @@
+import { getCurrentInstance, inject, onUnmounted } from 'vue';
+
+export const useParent = (key: symbol) => {
+  const parent = inject<any>(key, null);
+  if (parent) {
+    const instance = getCurrentInstance()!;
+    const { link, unlink } = parent;
+    link(instance);
+    onUnmounted(() => {
+      unlink(instance);
+    });
+  }
+  return { parent };
+};
